+++ src/platform.ts
@@ -63,12 +63,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
   public readonly eve: any;
   public readonly webhookEventHandler: { [x: string]: (context: { [x: string]: any }) => void } = {};
 
-  constructor(log: Logging, config: SwitchBotPlatformConfig, api: API) {
+  constructor(
+    log: Logging,
+    config: SwitchBotPlatformConfig,
+    api: API,
+  ) {
     this.accessories = [];
     this.api = api;
     this.log = log;
     // only load if configured
-    if (!this.config) {
+    if (!config) {
       return;
     }
 
@@ -77,23 +81,10 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
       platform: 'SwitchBotPlatform',
       name: config.name,
       credentials: config.credentials as object,
-      url: config.url as string,
-      port: config.port as number,
-      hostname: config.hostname as string,
-      protocol: config.protocol as string,
-      verifyTLS: config.verifyTLS as boolean,
-      logging: config.logging as string,
+      options: config.options as object,
     };
-    this.logs();
+    this.logType();
     this.debugLog(`Finished initializing platform: ${config.name}`);
-
-    // HOOBS notice
-    if (__dirname.includes('hoobs')) {
-      this.warnLog(
-        'This plugin has not been tested under HOOBS, it is highly recommended that you switch to Homebridge: ' +
-        'https://tinyurl.com/HOOBS2Homebridge',
-      );
-    }
 
     // verify the config
     try {
@@ -325,6 +316,8 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
    * Verify the config passed to the plugin is valid
    */
   async verifyConfig() {
+    this.debugLog('Verifying Config');
+    this.config = this.config || {};
     this.config.options = this.config.options || {};
 
     const platformConfig = {};
@@ -670,6 +663,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         this.createContact(device);
         break;
       case 'Curtain':
+      case 'Curtain3':
         this.debugLog(`Discovered ${device.deviceType} ${device.deviceName}: ${device.deviceId}`);
         this.createCurtain(device);
         break;
@@ -2313,16 +2307,15 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
   }
 
   async registerCurtains(device: device & devicesConfig) {
-    if (device.deviceType === 'Curtain') {
-      this.debugWarnLog(
-        `deviceName: ${device.deviceName} deviceId: ${device.deviceId}, curtainDevicesIds: ${device.curtainDevicesIds}, master: ` +
-        `${device.master}, group: ${device.group}, disable_group: ${device.curtain?.disable_group}, connectionType: ${device.connectionType}`,
-      );
-    } else {
-      this.debugWarnLog(
-        `deviceName: ${device.deviceName} deviceId: ${device.deviceId}, blindTiltDevicesIds: ${device.blindTiltDevicesIds}, master: ` +
-        `${device.master}, group: ${device.group}, disable_group: ${device.curtain?.disable_group}, connectionType: ${device.connectionType}`,
-      );
+    switch (device.deviceType) {
+      case 'Curtain':
+      case 'Curtain3':
+        this.debugWarnLog(`deviceName: ${device.deviceName} deviceId: ${device.deviceId}, curtainDevicesIds: ${device.curtainDevicesIds}, master: ` +
+          `${device.master}, group: ${device.group}, disable_group: ${device.curtain?.disable_group}, connectionType: ${device.connectionType}`);
+        break;
+      default:
+        this.debugWarnLog(`deviceName: ${device.deviceName} deviceId: ${device.deviceId}, blindTiltDevicesIds: ${device.blindTiltDevicesIds}, master:`
+        + ` ${device.master}, group: ${device.group}, disable_group: ${device.curtain?.disable_group}, connectionType: ${device.connectionType}`);
     }
 
     let registerCurtain: boolean;
@@ -2383,12 +2376,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     device.connectionType = await this.connectionType(device);
     let registerDevice: boolean;
     if (!device.hide_device && device.connectionType === 'BLE/OpenAPI') {
-      if (device.deviceType === 'Curtain' || device.deviceType === 'Blind Tilt') {
-        registerDevice = await this.registerCurtains(device);
-        this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
-      } else {
-        registerDevice = true;
-        this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
+      switch (device.deviceType) {
+        case 'Curtain':
+        case 'Curtain3':
+        case 'Blind Tilt':
+          registerDevice = await this.registerCurtains(device);
+          this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
+          break;
+        default:
+          registerDevice = true;
+          this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
       }
       if (registerDevice === true) {
         this.debugWarnLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will display in HomeKit`);
@@ -2396,12 +2393,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         this.debugErrorLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will not display in HomeKit`);
       }
     } else if (!device.hide_device && device.deviceId && device.configDeviceType && device.configDeviceName && device.connectionType === 'BLE') {
-      if (device.deviceType === 'Curtain' || device.deviceType === 'Blind Tilt') {
-        registerDevice = await this.registerCurtains(device);
-        this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
-      } else {
-        registerDevice = true;
-        this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
+      switch (device.deviceType) {
+        case 'Curtain':
+        case 'Curtain3':
+        case 'Blind Tilt':
+          registerDevice = await this.registerCurtains(device);
+          this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
+          break;
+        default:
+          registerDevice = true;
+          this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
       }
       if (registerDevice === true) {
         this.debugWarnLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will display in HomeKit`);
@@ -2409,12 +2410,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         this.debugErrorLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will not display in HomeKit`);
       }
     } else if (!device.hide_device && device.connectionType === 'OpenAPI') {
-      if (device.deviceType === 'Curtain' || device.deviceType === 'Blind Tilt') {
-        registerDevice = await this.registerCurtains(device);
-        this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
-      } else {
-        registerDevice = true;
-        this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
+      switch (device.deviceType) {
+        case 'Curtain':
+        case 'Curtain3':
+        case 'Blind Tilt':
+          registerDevice = await this.registerCurtains(device);
+          this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
+          break;
+        default:
+          registerDevice = true;
+          this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
       }
       if (registerDevice === true) {
         this.debugWarnLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will display in HomeKit`);
@@ -2422,12 +2427,16 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
         this.debugErrorLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will not display in HomeKit`);
       }
     } else if (!device.hide_device && device.connectionType === 'Disabled') {
-      if (device.deviceType === 'Curtain' || device.deviceType === 'Blind Tilt') {
-        registerDevice = await this.registerCurtains(device);
-        this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
-      } else {
-        registerDevice = true;
-        this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
+      switch (device.deviceType) {
+        case 'Curtain':
+        case 'Curtain3':
+        case 'Blind Tilt':
+          registerDevice = await this.registerCurtains(device);
+          this.debugWarnLog(`Device: ${device.deviceName} ${device.deviceType} registerDevice: ${registerDevice}`);
+          break;
+        default:
+          registerDevice = true;
+          this.debugWarnLog(`Device: ${device.deviceName} registerDevice: ${registerDevice}`);
       }
       this.debugWarnLog(`Device: ${device.deviceName} connectionType: ${device.connectionType}, will continue to display in HomeKit`);
     } else if (!device.connectionType && !device.hide_device) {
@@ -2529,7 +2538,7 @@ export class SwitchBotPlatform implements DynamicPlatformPlugin {
     return switchbot;
   }
 
-  logs() {
+  logType() {
     this.debugMode = process.argv.includes('-D') || process.argv.includes('--debug');
     if (this.config.options?.logging === 'debug' || this.config.options?.logging === 'standard' || this.config.options?.logging === 'none') {
       this.platformLogging = this.config.options!.logging;